auth for recipe delete

+++ jerk-chicken/src/main/java/com/jerk/chicken/controllers/RecipeController.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,6 +18,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.jerk.chicken.models.Recipe;
+import com.jerk.chicken.models.Role;
 import com.jerk.chicken.models.dto.ComplexRecipeDTO;
 import com.jerk.chicken.repositories.RecipeRepository;
 import com.jerk.chicken.services.RecipeService;
@@ -33,7 +35,7 @@ public class RecipeController {
 
 	@Autowired
 	RecipeRepository recipeRepo;
-	
+
 	@Autowired
 	JwtValidate jwt;
 
@@ -51,21 +53,27 @@ public class RecipeController {
 	@ResponseStatus(HttpStatus.CREATED)
 	public ComplexRecipeDTO addRecipe(@RequestBody ComplexRecipeDTO r, @RequestHeader("x-access-token") String token) {
 		UserData user = jwt.getUserData(token);
-		return rs.saveRecipe(r,user.getId());
+		return rs.saveRecipe(r, user.getId());
 	}
 
 	@DeleteMapping("{id}")
 	@ResponseStatus(HttpStatus.OK)
-	public void deleteRecipe(@PathVariable("id")int id) {
-		Recipe r = new Recipe();
-		r.setId(id);
-		rs.deleteRecipe(r);
+	public ResponseEntity<Boolean> deleteRecipe(@PathVariable("id") int id, @RequestHeader("x-access-token") String token) {
+		UserData user = jwt.getUserData(token);
+		if (user.getRoles().contains(new Role(1, "admin")) || (getRecipeById(id).getOwnerId() == user.getId())) {
+			Recipe r = new Recipe();
+			r.setId(id);
+			rs.deleteRecipe(r);
+			return new ResponseEntity<>(true, HttpStatus.OK);
+		}
+		return new ResponseEntity<>(false, HttpStatus.UNAUTHORIZED);
 	}
 
 	@PutMapping
 	@ResponseStatus(HttpStatus.OK)
-	public ComplexRecipeDTO updateRecipe(@RequestBody ComplexRecipeDTO r, @RequestHeader("x-access-token") String token) {
+	public ComplexRecipeDTO updateRecipe(@RequestBody ComplexRecipeDTO r,
+			@RequestHeader("x-access-token") String token) {
 		UserData user = jwt.getUserData(token);
-		return rs.saveRecipe(r,user.getId());
+		return rs.saveRecipe(r, user.getId());
 	}
 }
